Refactor useData to retrieve mock data through a promise.
