+++ .gitignore
@@ -22,3 +22,5 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 package-lock.json
+.env.development
+.env.production
