+++ src/hooks/useData.js
@@ -1,29 +1,37 @@
 import { useState, useEffect } from "react";
 
-const useData = (fileName, mapFunction) => {
+const useData = (service, fileName, mapFunction) => {
     const [parsedData, setParsedData] = useState([]);
    
     useEffect(() => {
-        const abortController = new AbortController();
+        const controller = new AbortController();
 
-        fetch(`wa-react-2021-q4-internal-capstone-project/${fileName}`, {
-            headers : { 
-              'Content-Type': 'application/json',
-              'Accept': 'application/json'
-             }
-           })
-            .then(response => response.json())
-            .then((data) => {
-                const results = data.results;
-        
-                var productCategoriesConverted = results.map(mapFunction);
-                setParsedData(productCategoriesConverted);
-            })
-            .catch(error => {
+        service(fileName, mapFunction, controller.signal)
+            .then((response) => {
+                setParsedData(response);
+            }).catch(error => {
                 console.error('There has been a problem with your fetch operation:', error);
             });
 
-        return () => abortController.abort();
+        // fetch(`wa-react-2021-q4-internal-capstone-project/${fileName}`, {
+        //     headers : { 
+        //       'Content-Type': 'application/json',
+        //       'Accept': 'application/json'
+        //      },
+        //      signal: controller.signal
+        //    })
+        //     .then(response => response.json())
+        //     .then((data) => {
+        //         const results = data.results;
+        
+        //         var productCategoriesConverted = results.map(mapFunction);
+        //         setParsedData(productCategoriesConverted);
+        //     })
+        //     .catch(error => {
+        //         console.error('There has been a problem with your fetch operation:', error);
+        //     });
+
+        return () => controller.abort();
     }, []);
     
     return [parsedData];
