+++ src/hooks/useFeaturedBanners.js
@@ -1,10 +1,11 @@
-import useData from 'hooks/useData';
+import useMockData from 'hooks/useData';
+import { devSearch } from 'utils/DataRetrievalService';
 
 const useFeaturedBanners = () => {
     const fileName = 'featured-banners.json';
     const mapFunction = ({id, data: { main_image: { url } }}) => { return { id, url } };
 
-    const [productCategories] = useData(fileName, mapFunction);
+    const [productCategories] = useMockData(devSearch, fileName, mapFunction);
     return [productCategories];
 };
